categories/joy/joy-4m-bardendy/solution/requirements.txt
Pillow==9.5.0
pwntools==4.9.0
pyzbar==0.1.9
tqdm==4.65.0
